# alphadash/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AlphaDash",
  "description": "Beginner strategy-agent product \u2014 paper-trading first.",
  "version": "0.0.1"
 },
 "paths": {
...
  "/account": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Account Info",
    "operationId": "account_info_account_get",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/account/limits": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Account Limits",
    "operationId": "account_limits_account_limits_get",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LimitsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/account/pause": {
   "post": {
    "tags": [
     "portfolio"
    ],
    "summary": "Pause Account",
    "operationId": "pause_account_account_pause_post",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/account/resume": {
   "post": {
    "tags": [
     "portfolio"
    ],
    "summary": "Resume Account",
    "operationId": "resume_account_account_resume_post",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portfolio": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Portfolio",
    "operationId": "portfolio_portfolio_get",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PortfolioOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portfolio/performance": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Performance",
    "operationId": "performance_portfolio_performance_get",
    "parameters": [
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 365,
       "minimum": 7,
       "default": 90,
       "title": "Days"
      }
     },
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerformanceOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/quotes/{symbol}": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Quote",
    "operationId": "quote_quotes__symbol__get",
    "parameters": [
     {
      "name": "symbol",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Symbol"
      }
     },
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QuoteOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AccountOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "base_currency": {
      "type": "string",
      "title": "Base Currency"
     },
     "starting_equity": {
      "type": "string",
      "title": "Starting Equity"
     },
     "paused": {
      "type": "boolean",
      "title": "Paused"
     },
     "trading_mode": {
      "type": "string",
      "title": "Trading Mode"
     }
    },
    "type": "object",
    "required": [
     "id",
     "mode",
     "base_currency",
     "starting_equity",
     "paused",
     "trading_mode"
    ],
    "title": "AccountOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LimitsOut": {
    "properties": {
     "max_position_pct": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Position Pct"
     },
     "max_asset_class_pct": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Max Asset Class Pct"
     },
     "max_trades_per_week": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Trades Per Week"
     },
     "cash_floor_pct": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cash Floor Pct"
     },
     "per_suggestion_max_pct": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Per Suggestion Max Pct"
     },
     "drawdown_pause_pct": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Drawdown Pause Pct"
     }
    },
    "type": "object",
    "required": [
     "max_position_pct",
     "max_asset_class_pct",
     "max_trades_per_week",
     "cash_floor_pct",
     "per_suggestion_max_pct",
     "drawdown_pause_pct"
    ],
    "title": "LimitsOut"
   },
...
   "PerformanceOut": {
    "properties": {
     "points": {
      "items": {
       "$ref": "#/components/schemas/PerformancePointOut"
      },
      "type": "array",
      "title": "Points"
     },
     "return_pct": {
      "type": "number",
      "title": "Return Pct"
     },
     "benchmark_return_pct": {
      "type": "number",
      "title": "Benchmark Return Pct"
     },
     "max_drawdown_pct": {
      "type": "number",
      "title": "Max Drawdown Pct"
     },
     "current_drawdown_pct": {
      "type": "number",
      "title": "Current Drawdown Pct"
     },
     "benchmark_symbol": {
      "type": "string",
      "title": "Benchmark Symbol"
     }
    },
    "type": "object",
    "required": [
     "points",
     "return_pct",
     "benchmark_return_pct",
     "max_drawdown_pct",
     "current_drawdown_pct",
     "benchmark_symbol"
    ],
    "title": "PerformanceOut"
   },
   "PerformancePointOut": {
    "properties": {
     "day": {
      "type": "string",
      "title": "Day"
     },
     "equity": {
      "type": "string",
      "title": "Equity"
     },
     "benchmark_equity": {
      "type": "string",
      "title": "Benchmark Equity"
     }
    },
    "type": "object",
    "required": [
     "day",
     "equity",
     "benchmark_equity"
    ],
    "title": "PerformancePointOut"
   },
   "PortfolioOut": {
    "properties": {
     "equity": {
      "type": "string",
      "title": "Equity"
     },
     "cash": {
      "type": "string",
      "title": "Cash"
     },
     "positions": {
      "items": {
       "$ref": "#/components/schemas/PositionOut"
      },
      "type": "array",
      "title": "Positions"
     },
     "allocation_pct": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Allocation Pct"
     }
    },
    "type": "object",
    "required": [
     "equity",
     "cash",
     "positions",
     "allocation_pct"
    ],
    "title": "PortfolioOut"
   },
   "PositionOut": {
    "properties": {
     "symbol": {
      "type": "string",
      "title": "Symbol"
     },
     "asset_class": {
      "type": "string",
      "title": "Asset Class"
     },
     "quantity": {
      "type": "string",
      "title": "Quantity"
     },
     "avg_cost": {
      "type": "string",
      "title": "Avg Cost"
     },
     "market_value": {
      "type": "string",
      "title": "Market Value"
     },
     "unrealized_pl": {
      "type": "string",
      "title": "Unrealized Pl"
     },
     "allocation_pct": {
      "type": "number",
      "title": "Allocation Pct"
     }
    },
    "type": "object",
    "required": [
     "symbol",
     "asset_class",
     "quantity",
     "avg_cost",
     "market_value",
     "unrealized_pl",
     "allocation_pct"
    ],
    "title": "PositionOut"
   },
   "QuoteOut": {
    "properties": {
     "symbol": {
      "type": "string",
      "title": "Symbol"
     },
     "price": {
      "type": "string",
      "title": "Price"
     },
     "as_of": {
      "type": "string",
      "title": "As Of"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "is_stale": {
      "type": "boolean",
      "title": "Is Stale"
     }
    },
    "type": "object",
    "required": [
     "symbol",
     "price",
     "as_of",
     "source",
     "is_stale"
    ],
    "title": "QuoteOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}